+++ redvox/common/reader_session_model.py
@@ -3,9 +3,9 @@ This module describes methods to store session models from different sources.
 This information can be utilized by users as they see fit
 """
 
-from typing import List, Optional, Callable
+from typing import List, Optional
 
-from redvox.common.session_model import SessionModel
+from redvox.common.session_model import SessionModel, LocalSessionModels
 from redvox.common.errors import RedVoxExceptions
 from redvox.cloud.client import cloud_client
 from redvox.cloud.session_model_api import SessionModelsResp, Session, DynamicSession
@@ -23,28 +23,24 @@ class ModelsContainer:
     Properties:
         cloud_models: SessionModelsResp that contains all the cloud models
 
-        local_models: List[redvox.SessionModel] that contains all the locally created models
+        local_models: LocalSessionModels that contains all the locally created models
     """
 
-    def __init__(self, cloud: Optional[SessionModelsResp] = None, local: Optional[List[SessionModel]] = None):
+    def __init__(self, cloud: Optional[SessionModelsResp] = None, local: Optional[LocalSessionModels] = None):
         """
         initialize the container.
-
         :param cloud: SessionModelsResp containing all cloud session models
         :param local: list of local session models
         """
         self.cloud_models: Optional[SessionModelsResp] = cloud
-        self.local_models: Optional[List[SessionModel]] = local
+        self.local_models: Optional[LocalSessionModels] = local
         self._errors: RedVoxExceptions = RedVoxExceptions("ModelsContainer")
 
     def __repr__(self):
         """
         :return: representation of ModelsContainer
         """
-        return (
-            f"cloud_models: {self.cloud_models.__repr__()}, "
-            f"local_models: {[f.__repr__() for f in self.local_models]}"
-        )
+        return f"cloud_models: {self.cloud_models.__repr__()}, " f"local_models: {self.local_models.__repr__()}"
 
     def as_dict(self) -> dict:
         """
@@ -52,7 +48,7 @@ class ModelsContainer:
         """
         return {
             "cloud_models": self.cloud_models.to_dict() if self.cloud_models else None,
-            "local_models": [f.as_dict() for f in self.local_models] if self.local_models else None,
+            "local_models": self.local_models.as_dict() if self.local_models else None,
             "errors": self._errors.as_dict(),
         }
 
@@ -64,22 +60,33 @@ class ModelsContainer:
         """
         result = ModelsContainer(
             SessionModelsResp.from_dict(in_dict["cloud_models"]) if "cloud_models" in in_dict.keys() else None,
-            [SessionModel.from_dict(f) for f in in_dict["local_models"]] if "local_models" in in_dict.keys() else None,
+            LocalSessionModels.from_dict(in_dict["local_models"]) if "local_models" in in_dict.keys() else None,
         )
-        result._errors = in_dict["errors"]
+        if "errors" in in_dict.keys():
+            result._errors = in_dict["errors"]
         return result
 
     @staticmethod
-    def _model_operation(model: Session, func: Callable):
+    def _key_check(model: Session, station_id: str, uuid: Optional[str], start_date: Optional[int]):
         """
-        :param model: Session model to operate on
-        :param func: function to perform; takes model as a parameter
-        :return: output of the function
+        check a key using components instead of the whole key.
+        :param model: Session model to compare with
+        :param station_id: station ID to check
+        :param uuid: uuid to check
+        :param start_date: station start date to check
+        :return: True if all passed components are the same as the model's.
         """
-        return func(model)
+        return (
+            (model.id == station_id)
+            and (not uuid and not start_date)
+            or (uuid and model.uuid == uuid and not start_date)
+            or (start_date and model.start_ts == start_date and not uuid)
+            or (uuid and model.uuid == uuid and start_date and model.start_ts == start_date)
+        )
 
     def get_model(self, session_key: str) -> Optional[Session]:
         """
+        Prioritizes returning cloud models over local models.
         :param session_key: the key for the session.  format is: "{STATION_ID}:{STATION_UUID}:{STATION_START_DATE}"
             where STATION_START_DATE is the integer start date of the station in microseconds since epoch UTC.
         :return: cloud session model matching the key or None
@@ -89,15 +96,16 @@ class ModelsContainer:
                 if session_key == n.session_key():
                     return n
         if self.local_models:
-            for n in self.local_models:
+            for n in self.local_models.sessions:
                 if session_key == n.cloud_session.session_key():
                     return n.cloud_session
         return None
 
-    def get_model_by_key(
+    def get_model_by_partial_key(
         self, station_id: str, uuid: Optional[str] = None, start_date: Optional[int] = None
     ) -> Optional[Session]:
         """
+        Prioritizes returning cloud models over local models.
         :param station_id: id of the station to get
         :param uuid: uuid of station to get.  if None, gets the first station that matches other parameters.  Default
                 None
@@ -106,22 +114,14 @@ class ModelsContainer:
         :return: First session that matches parameters or None
         """
         # if uuid or start date given, check that it matches, if both given, check if both match
-        key_check_func = (
-            lambda x: (not uuid and not start_date)
-            or (uuid and x.uuid == uuid and not start_date)
-            or (start_date and x.start_ts == start_date and not uuid)
-            or (uuid and x.uuid == uuid and start_date and x.start_ts == start_date)
-        )
         if self.cloud_models:
             for n in self.cloud_models.sessions:
-                if n.id == station_id:
-                    if self._model_operation(n, key_check_func):
-                        return n
+                if self._key_check(n, station_id, uuid, start_date):
+                    return n
         if self.local_models:
-            for n in self.local_models:
-                if n.cloud_session.id == station_id:
-                    if self._model_operation(n.cloud_session, key_check_func):
-                        return n.cloud_session
+            for n in self.local_models.sessions:
+                if self._key_check(n.cloud_session, station_id, uuid, start_date):
+                    return n.cloud_session
         return None
 
     def get_all_models(self) -> List[Session]:
@@ -133,12 +133,13 @@ class ModelsContainer:
             for n in self.cloud_models.sessions:
                 models.append(n)
         if self.local_models:
-            for n in self.local_models:
+            for n in self.local_models.sessions:
                 models.append(n.cloud_session)
         return models
 
     def get_dynamic_session(self, key: str) -> Optional[DynamicSession]:
         """
+        Prioritizes returning cloud models over local models.
         :param key: key to the dynamic session, formatted as: ID:UUID:SESSION_START_DATE:DYNAMIC_START:DYNAMIC_END
                     where START and END values times as microseconds since epoch UTC.
         :return: DynamicSession matching the key or None
@@ -155,7 +156,7 @@ class ModelsContainer:
             pass
         finally:
             if not dynamic_session:
-                for n in self.local_models:
+                for n in self.local_models.sessions:
                     for d, s in n.dynamic_sessions.items():
                         if d == key:
                             return s
@@ -170,7 +171,7 @@ class ModelsContainer:
 
     def add_cloud_session(self, new_session: Session):
         """
-        Add a cloud Session to the container.
+        Add a cloud Session to the container.  Does nothing if you try to add an existing key
         :param new_session: session to add
         """
         if self.cloud_models:
@@ -221,16 +222,17 @@ class ModelsContainer:
 
     def add_local_session(self, new_session: SessionModel):
         """
-        Add a local SessionModel to the container.
+        Add a local SessionModel to the container.  Does nothing if you try to add an existing key
         :param new_session: session to add
         """
+        new_session_key = new_session.cloud_session.session_key()
         if self.local_models:
-            if self.get_model(new_session.cloud_session.session_key()):
-                self._errors.append(f"Attempted to add existing key {new_session.cloud_session.session_key()}")
-            else:
-                self.local_models.append(new_session)
+            if self.get_model(new_session_key):
+                self._errors.append(f"Attempted to add existing key {new_session_key}")
+                return
         else:
-            self.local_models = [new_session]
+            self.local_models = LocalSessionModels()
+        self.local_models.sessions.append(new_session)
 
     def errors(self) -> RedVoxExceptions:
         """
@@ -247,7 +249,7 @@ class ModelsContainer:
             for k in self.cloud_models.sessions:
                 keys.append(k.session_key())
         if self.local_models:
-            for k in self.local_models:
+            for k in self.local_models.sessions:
                 keys.append(k.cloud_session.session_key())
         return keys
 
@@ -255,13 +257,13 @@ class ModelsContainer:
         """
         :return: all station ids in the models
         """
-        keys = []
+        ids = []
         if self.cloud_models:
             for k in self.cloud_models.sessions:
-                if k.id not in keys:
-                    keys.append(k.id)
+                if k.id not in ids:
+                    ids.append(k.id)
         if self.local_models:
-            for k in self.local_models:
-                if k.cloud_session.id not in keys:
-                    keys.append(k.cloud_session.id)
-        return keys
+            for k in self.local_models.sessions:
+                if k.cloud_session.id not in ids:
+                    ids.append(k.cloud_session.id)
+        return ids